+++ .github/CHANGELOG.md
@@ -4,6 +4,9 @@
  * Add probability support.
  [(#11)](https://github.com/PennyLaneAI/pennylane-lightning-kokkos/pull/11)
  
+ * Add sample generation support.
+  [(#9)](https://github.com/PennyLaneAI/pennylane-lightning-kokkos/pull/9)
+
 ### Breaking changes
 
 ### Improvements